requirements.txt
coverage==4.4.2
deprecated==1.1.0
flask==1.0.2
iso8601==0.1.12
nose==1.3.7
nose-timer==0.7.1
requests==2.19.1
gunicorn==19.9.0
jsonschema==2.6.0
flask-restplus==0.11.0
pyjwt==1.6.4
croniter==0.3.25
python-crontab==2.3.5
httpretty==0.9.5
pyopenssl==18.0.0
flask-cors==3.0.6
pymongo==3.7.2
mongomock==3.13.0
ipython
